lunia-bot/requirements.txt
python-telegram-bot 
python-dotenv 